+++ examples/src/lib.rs
@@ -1,4 +1,4 @@
-use fusen_rs::fusen_procedural_macro::{asset, fusen_trait};
+use fusen_rs::fusen_procedural_macro::fusen_trait;
 use serde::{Deserialize, Serialize};
 pub mod handler;
 
@@ -18,16 +18,16 @@ pub trait DemoService {
 
     async fn sayHello(&self, name: String) -> String;
 
-    #[asset(path = "/sayHelloV2-http")]
+    #[fusen_rs::fusen_procedural_macro::asset(path = "/sayHelloV2-http")]
     async fn sayHelloV2(&self, name: RequestDto) -> ResponseDto;
 
-    #[asset(path = "/divide", method = GET)]
+    #[fusen_rs::fusen_procedural_macro::asset(path = "/divide", method = GET)]
     async fn divideV2(&self, a: i32, b: i32) -> String;
 }
 
 #[fusen_trait(group = "v1", version = "1.0")]
-#[asset(path = "/dome")]
+#[fusen_rs::fusen_procedural_macro::asset(path = "/dome")]
 pub trait DemoServiceV2 {
-    #[asset(path = "/sayHelloV3-http")]
+    #[fusen_rs::fusen_procedural_macro::asset(path = "/sayHelloV3-http")]
     async fn sayHelloV3(&self, name: RequestDto) -> ResponseDto;
 }